*: move properties and compact_listener into engine_rocks
