+++ components/engine_rocks/Cargo.toml
@@ -39,3 +39,4 @@ package = "rocksdb"
 
 [dev-dependencies]
 tempfile = "3.0"
+rand = "0.6.5"
